faster translate algorithm of FixedTranslator. After find sync loop in line 54 we know that we either didn't find another sync, or found one and looking for a full message.
in the first case, it is enough to keep just the last sync_size bytes of input_so_far, because we are looking for sync.
in the second case, we can drop the sync and everything before it, and simply look for a message in the next calls to translate.